+++ kong/plugins/openid-connect/migrations/001_14_to_15.lua
@@ -1,10 +1,8 @@
 return {
   postgres = {
     up = [[
-
       -- Unique constraint on "issuer" already adds btree index
       DROP INDEX IF EXISTS "oic_issuers_idx";
-
     ]],
   },