+++ kong/plugins/openid-connect/migrations/002_200_to_210.lua
@@ -1,0 +1,19 @@
+return {
+  postgres = {
+    up = [[
+      CREATE TABLE IF NOT EXISTS "oic_jwks" (
+        "id"    UUID    PRIMARY KEY,
+        "jwks"  JSONB
+      );
+    ]],
+  },
+
+  cassandra = {
+    up = [[
+      CREATE TABLE IF NOT EXISTS oic_jwks (
+        id   uuid  PRIMARY KEY,
+        jwks text
+      );
+    ]],
+  },
+}