+++ kong/plugins/openid-connect/migrations/init.lua
@@ -1,4 +1,5 @@
 return {
   "000_base_openid_connect",
   "001_14_to_15",
+  "002_200_to_210",
 }